+++ app/dashboard/admin/QuizCreation.tsx
@@ -4,7 +4,6 @@ import { createClient } from "@supabase/supabase-js";
 import * as yup from "yup";
 import { useState } from "react";
 import { v4 as uuidv4 } from "uuid";
-import EditQuizzes from "@/components/EditQuizzes";
 
 const supabase = createClient(
   process.env.NEXT_PUBLIC_SUPABASE_URL!,
@@ -223,7 +222,6 @@ export default function QuizCreation() {
           </Form>
         )}
       </Formik>
-      <EditQuizzes/>
     </div>
   );
 }
